refactors code of saveOrUpdate method

+++ memoryorm/src/main/java/no/hyper/memoryorm/Memory.java
@@ -70,9 +70,9 @@ public class Memory {
         return result;
     }
 
-    public <T> int update(T entity) {
+    public <T> long update(T entity) {
         db.open();
-        int result = operationHelper.update(entity);
+        long result = operationHelper.update(entity);
         db.close();
         return result;
     }
